requirements.txt
numpy~=2.1.3
requests~=2.32.3
Flask~=3.0.3
scikit-learn~=1.5.2
python-dotenv~=1.0.1